Fix: Response 분리
